style : trip 엔티티 수정
컬럼 타입 Float 수정

+++ src/main/java/com/nadeul/ndj/entity/Trip.java
@@ -21,9 +21,9 @@ public class Trip {
   
   private Integer memId;
   
-  private Integer posX;
+  private Float posX;
   
-  private Integer posY;
+  private Float posY;
